mapped a couple pool-related test scripts
