+++ Makefile
@@ -71,6 +71,10 @@ submodules:
 test-pool-upgrade:
 	make TEST_TARGET=pool-upgrade.spec.ts test
 
+.PHONY: test-pool-edge
+test-pool-edge:
+	make TEST_TARGET=pool-edge.spec.ts test
+
 .PHONY: test-ntoken
 test-ntoken:
 	make TEST_TARGET=_xtoken_ntoken.spec.ts test
@@ -114,6 +118,10 @@ test-configurator:
 .PHONY: test-pausable-reserve
 test-pausable-reserve:
 	make TEST_TARGET=pausable-reserve.spec.ts test
+
+.PHONY: test-rescue-tokens
+test-rescue-tokens:
+	make TEST_TARGET=_pool_parameters_rescue_tokens.spec.ts test	
 
 .PHONY: test-upgradeability
 test-upgradeability:
