+++ test-suites/_base_upgradeability.spec.ts
@@ -17,15 +17,23 @@ import {
   deployMockPToken,
   deployMockReentrantInitializableImple,
   deployMockVariableDebtToken,
+  deployPoolCoreLibraries,
 } from "../deploy/helpers/contracts-deployments";
-import {getEthersSigners} from "../deploy/helpers/contracts-helpers";
+import {
+  getEthersSigners,
+  getFunctionSignatures,
+} from "../deploy/helpers/contracts-helpers";
 import {
   InitializableImmutableAdminUpgradeabilityProxy,
   InitializableImmutableAdminUpgradeabilityProxy__factory,
+  PoolCoreV2__factory,
+  PoolCore__factory,
 } from "../types";
 import {ProtocolErrors} from "../deploy/helpers/types";
 import {loadFixture} from "@nomicfoundation/hardhat-network-helpers";
 import {testEnvFixture} from "./helpers/setup-env";
+import {liquidateAndValidateReverted} from "./helpers/validated-steps";
+import {waitForTx} from "../deploy/helpers/misc-utils";
 
 describe("Upgradeability", () => {
   let testEnv: TestEnv;
@@ -618,6 +626,118 @@ describe("Upgradeability", () => {
       expect(tokenName).to.be.eq(
         "ParaSpace variable debt bearing DAI updated",
         "Invalid token name"
+      );
+    });
+  });
+
+  context("Pool Upgrade", () => {
+    let testEnv: TestEnv;
+
+    beforeEach(async () => {
+      testEnv = await loadFixture(testEnvFixture);
+    });
+
+    it("TC-upgradeability-32 Disable liquidation by remove liquidateERC20 in current pool", async () => {
+      const {
+        addressesProvider,
+        weth,
+        usdc,
+        users: [borrower, liquidator],
+      } = testEnv;
+
+      await liquidateAndValidateReverted(
+        weth,
+        usdc,
+        "1000",
+        liquidator,
+        borrower,
+        false,
+        //before upgrade only protocolError
+        ProtocolErrors.HEALTH_FACTOR_NOT_BELOW_THRESHOLD
+      );
+
+      const liquidateERC20Signature = getFunctionSignatures(
+        PoolCore__factory.abi
+      ).filter((s) => s.name.includes("liquidateERC20"))[0].signature;
+
+      await waitForTx(
+        await addressesProvider.updatePoolImpl(
+          [
+            {
+              implAddress: ZERO_ADDRESS,
+              action: 2, //remove function
+              functionSelectors: [liquidateERC20Signature],
+            },
+          ],
+          ZERO_ADDRESS,
+          "0x"
+        )
+      );
+
+      await liquidateAndValidateReverted(
+        weth,
+        usdc,
+        "1000",
+        liquidator,
+        borrower,
+        false,
+        //after upgrade error as "function does not exist"
+        "ParaProxy: Function does not exist"
+      );
+    });
+
+    it("TC-upgradeability-33: Disable liquidation by upgrading to a new pool contract which will revert liquidateERC20", async () => {
+      const {
+        addressesProvider,
+        weth,
+        usdc,
+        users: [borrower, liquidator],
+      } = testEnv;
+
+      await liquidateAndValidateReverted(
+        weth,
+        usdc,
+        "1000",
+        liquidator,
+        borrower,
+        false,
+        ProtocolErrors.HEALTH_FACTOR_NOT_BELOW_THRESHOLD
+      );
+
+      const poolCoreV2Selectors = getFunctionSignatures(
+        PoolCoreV2__factory.abi
+      ).map((s) => s.signature);
+
+      const coreLibraries = await deployPoolCoreLibraries();
+
+      const poolCoreV2 = await new PoolCoreV2__factory(
+        coreLibraries,
+        await getFirstSigner()
+      ).deploy(addressesProvider.address);
+
+      await waitForTx(
+        await addressesProvider.updatePoolImpl(
+          [
+            {
+              implAddress: poolCoreV2.address,
+              action: 1, //replace
+              functionSelectors: poolCoreV2Selectors,
+            },
+          ],
+          ZERO_ADDRESS,
+          "0x"
+        )
+      );
+
+      await liquidateAndValidateReverted(
+        weth,
+        usdc,
+        "1000",
+        liquidator,
+        borrower,
+        false,
+        //after upgrading to a new contract error as "emergency disable call"
+        ProtocolErrors.EMEGENCY_DISABLE_CALL
       );
     });
   });
