+++ test-suites/pool-edge.spec.ts
@@ -5,10 +5,13 @@ import {MAX_UINT_AMOUNT, ZERO_ADDRESS} from "../deploy/helpers/constants";
 import {
   deployReserveAuctionStrategy,
   deployMintableERC20,
+  deployPoolComponents,
+  deployPoolCoreLibraries,
 } from "../deploy/helpers/contracts-deployments";
 import {eContractid, ProtocolErrors} from "../deploy/helpers/types";
 import {
   MockReserveInterestRateStrategy__factory,
+  PoolCore__factory,
   PToken__factory,
   VariableDebtToken__factory,
 } from "../types";
@@ -32,6 +35,7 @@ describe("Pool: Edge cases", () => {
     NOT_CONTRACT,
     RESERVE_ALREADY_INITIALIZED,
     RESERVE_ALREADY_ADDED,
+    INVALID_ADDRESSES_PROVIDER,
   } = ProtocolErrors;
 
   const MAX_NUMBER_RESERVES = 128;
@@ -114,23 +118,22 @@ describe("Pool: Edge cases", () => {
   //   expect(userReserveDataAfter.healthFactor).to.be.eq(MAX_UINT_AMOUNT);
   // });
 
-  // it("Initialize fresh deployment with incorrect addresses provider (revert expected)", async () => {
-  //   const {
-  //     addressesProvider,
-  //     users: [deployer],
-  //   } = testEnv;
-  //
-  //   const NEW_POOL_IMPL_ARTIFACT = await deployPool(addressesProvider.address);
-  //
-  //   const freshPool = Pool__factory.connect(
-  //     NEW_POOL_IMPL_ARTIFACT.address,
-  //     deployer.signer
-  //   );
-  //
-  //   await expect(freshPool.initialize(deployer.address)).to.be.revertedWith(
-  //     INVALID_ADDRESSES_PROVIDER
-  //   );
-  // });
+  it("Initialize fresh deployment with incorrect addresses provider (revert expected)", async () => {
+    const {
+      addressesProvider,
+      users: [deployer],
+    } = testEnv;
+
+    const coreLibraries = await deployPoolCoreLibraries(false);
+    const poolCore = await new PoolCore__factory(
+      coreLibraries,
+      await getFirstSigner()
+    ).deploy(addressesProvider.address);
+
+    await expect(poolCore.initialize(deployer.address)).to.be.revertedWith(
+      INVALID_ADDRESSES_PROVIDER
+    );
+  });
 
   it("Check initialization", async () => {
     const {pool} = testEnv;
